rename to jsx, fix eslint
