--- frontend/src/contexts/AuthProvider.js
@@ -1,22 +1,0 @@
-import React, { createContext, useState } from 'react'
-
-export const AuthContext = createContext()
-
-const AuthProvider = ({ children }) => {
-    const [isAuthenticated, setIsAuthenticated] = useState(false)
-
-    const login = () => {
-        setIsAuthenticated(true)
-    }
-
-    const logout = () => {
-        setIsAuthenticated(false)
-    }
-
-    return (
-        <AuthContext.Provider value={{ isAuthenticated, login, logout }}>
-            {children}
-        </AuthContext.Provider>
-    )
-}
-export default AuthProvider
